pyproject.toml
[tool.poetry]
name = "optionscope"
version = "0.1.0"
description = "A production-grade options scanner/optimizer for evaluating and ranking options strategies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
yfinance = "^0.2.28"
scipy = "^1.11.0"
pydantic = "^2.4.0"
plotly = "^5.15.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[tool.poetry.scripts]
app = "streamlit run optionscope.py"
test = "pytest -q"
lint = "ruff check ."
typecheck = "mypy app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "B", "A"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
